add: settings page
