+++ web/src/components/organisms/Settings/General/index.tsx
@@ -1,0 +1,7 @@
+import SettingsMolecule from "@reearth-cms/components/molecules/Settings";
+
+const WorkspaceSettings: React.FC = () => {
+  return <SettingsMolecule />;
+};
+
+export default WorkspaceSettings;
